Maps update and fix error
file's change: adocao, animal_profile, criar_conta, criar_conta_ong, css\Ong_Profile, css\report, dashboard\controller\atualizarPerfil, dashboard\controller\updateReport, dashboard\css_dashboard\maps, dashboard\perfil, dashboard\visualizaReport, includes\modal.php, includes\nav, login, ong_profile, ongs, reportar

file's create: css\maps, dashboard\js\autocomplete, dashboard\js\gps, \dashboard\js\maps, dashboard\js\mapsReport, dashboard\js\nomeDaRua

+++ dashboard/controller/atualizarPerfil.php
@@ -113,11 +113,12 @@ if($_FILES['arquivo']['name'] != ""){
             if($linha = $img->fetch(PDO::FETCH_ASSOC)){
                 $name = $linha['ong_img']; //Se existir vai entrar na variavel $name
             }
-
-            if(file_exists("../../imgs/$name")) { //verificando se ela existe no diretorio
-                unlink("../../imgs/$name"); //Tirando a imgagem do diretorio
-            } 
-
+            if($name != "default_avatar.jpg"){
+                if(file_exists("../../imgs/$name")) { //verificando se ela existe no diretorio
+                    unlink("../../imgs/$name"); //Tirando a imgagem do diretorio
+                } 
+            }
+           
             //diretorio
             $uploaddir = "../../imgs/";
